services/openai_client.py
import os
from dotenv import load_dotenv
from openai import OpenAI

# ✅ Load .env explicitly
load_dotenv()

api_key = os.getenv("OPENAI_API_KEY")

if not api_key:
    raise RuntimeError("OPENAI_API_KEY not found. Check your .env file.")

client = OpenAI(api_key=api_key)


def get_chat_completion(messages, temperature=0.5):
    response = client.chat.completions.create(
        model="gpt-4o-mini",
        messages=messages,
        temperature=temperature
    )
    return response.choices[0].message.content
