requirements.txt
streamlit
openai
python-dotenv
pypdf
pydub
soundfile
